+++ hail/src/main/scala/is/hail/expr/types/physical/PTuple.scala
@@ -9,32 +9,6 @@ import is.hail.utils._
 
 case class PTupleField(index: Int, typ: PType)
 
-class CodePTuple(
-  val pType: PTuple,
-  val region: Code[Region],
-  val offset: Code[Long]
-) {
-  def apply[T](i: Int): Code[T] =
-    Region.loadIRIntermediate(pType.types(i))(
-      pType.loadField(offset, i)
-    ).asInstanceOf[Code[T]]
-
-  def isMissing(i: Int): Code[Boolean] = {
-    pType.isFieldMissing(offset, i)
-  }
-
-  private def withTypesAndIndices = (0 until pType.nFields).map(i => (pType.types(i), apply(i), i))
-
-  def withTypes = withTypesAndIndices.map(x => (x._1, x._2))
-
-  def missingnessPattern = (0 until pType.nFields).map(isMissing(_))
-
-  def values[T, U, V] = {
-    assert(pType.nFields == 3)
-    (apply[T](0), apply[U](1), apply[V](2))
-  }
-}
-
 object PTuple {
   def apply(required: Boolean, args: PType*): PTuple = PCanonicalTuple(args.iterator.zipWithIndex.map { case (t, i) => PTupleField(i, t)}.toFastIndexedSeq, required)
 
@@ -72,4 +46,30 @@ abstract class PTuple extends PBaseStruct {
   def copy(required: Boolean): PTuple
 
   def fieldIndex: Map[Int, Int]
+}
+
+class CodePTuple(
+  val pType: PTuple,
+  val region: Code[Region],
+  val offset: Code[Long]
+) {
+  def apply[T](i: Int): Code[T] =
+    Region.loadIRIntermediate(pType.types(i))(
+      pType.loadField(offset, i)
+    ).asInstanceOf[Code[T]]
+
+  def isMissing(i: Int): Code[Boolean] = {
+    pType.isFieldMissing(offset, i)
+  }
+
+  private def withTypesAndIndices = (0 until pType.nFields).map(i => (pType.types(i), apply(i), i))
+
+  def withTypes = withTypesAndIndices.map(x => (x._1, x._2))
+
+  def missingnessPattern = (0 until pType.nFields).map(isMissing(_))
+
+  def values[T, U, V] = {
+    assert(pType.nFields == 3)
+    (apply[T](0), apply[U](1), apply[V](2))
+  }
 }
